mitty-hns.12: Quality gate — fix bugs from code review + CodeRabbit
Production fixes:
- SSRF prevention: manual redirect following with metadata endpoint blocking (extract.py)
- Cache collision fix: include concept name in source hash (compiler.py)
- Guide timeout increased 4s→15s, block_type validation (generator.py)
- FK type alignment: block_id BigInteger→Integer (db.py, migration)
- DOCX MIME type added to extractable content types (fetcher.py)
- x-model two-way binding fix: :value+@input pattern (study_plan.html)
- Insert→upsert for guide persistence (generator.py, compiler.py)

Test fixes:
- Mock status_code for redirect-following tests (test_extract.py)
- WritableTable mock upsert method (test_generator.py)
- Corrected threshold test name (test_sources.py)

+++ mitty/db.py
@@ -691,7 +691,7 @@ study_block_guides = sa.Table(
     sa.Column("id", sa.BigInteger, primary_key=True, autoincrement=True),
     sa.Column(
         "block_id",
-        sa.BigInteger,
+        sa.Integer,
         sa.ForeignKey("study_blocks.id", ondelete="CASCADE"),
         nullable=False,
         unique=True,
@@ -722,7 +722,7 @@ block_artifacts = sa.Table(
     sa.Column("id", sa.BigInteger, primary_key=True, autoincrement=True),
     sa.Column(
         "block_id",
-        sa.BigInteger,
+        sa.Integer,
         sa.ForeignKey("study_blocks.id", ondelete="CASCADE"),
         nullable=False,
     ),
